+++ src/main/java/com/demo/spring/boot/SpringBootArtifactDemo/dto/CityDTO.java
@@ -6,13 +6,13 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class CityDTO {
-    private Integer id;
+    private Long id;
 
-    private int reference;
+    private Long reference;
     private String name;
     private String country;
 
-    public CityDTO(int reference, String name, String country) {
+    public CityDTO(Long reference, String name, String country) {
         this.reference = reference;
         this.name = name;
         this.country = country;
