+++ src/main/java/com/demo/spring/boot/SpringBootArtifactDemo/services/interfaces/TripService.java
@@ -3,7 +3,7 @@ package com.demo.spring.boot.SpringBootArtifactDemo.services.interfaces;
 import com.demo.spring.boot.SpringBootArtifactDemo.entities.Trip;
 
 public interface TripService {
-    Trip getTripById(int id);
+    Trip getTripById(Long id);
 
     Trip createTrip(Trip trip);
 }
